+++ CHANGES.md
@@ -1,5 +1,18 @@
 ## Changes in 1.4.2 (in development)
 
+* Enhanced spatial resampling in module `xcube.core.resampling` (#955): 
+    - Added optional keyword argument `interpolation` to function
+      `rectify_dataset()` with values `"nearest"`, `"triangular"`, 
+      and `"bilinear"` where `"triangular"` interpolates between 3 
+      and `"bilinear"` between 4 adjacent source pixels. 
+    - Function `rectify_dataset()` is now ~2 times faster by early 
+      detection of already transformed target pixels.      
+    - Added optional keyword argument `rectify_kwargs` to 
+      `resample_in_space()`. If given, it is spread into keyword arguments 
+      passed to the internal `rectify_dataset()` delegation, if any.
+    - Deprecated unused keyword argument `xy_var_names` of 
+      function `rectify_dataset()`.
+
 * Replace use of deprecated method in testing module. (#961)
 
 * Update dependencies to better match imports; remove the defaults channel;